Relocate bleu to integrations
